Rename NewTimeOfYearSelector -> TimeOfYearSelector
